Kobolds eating Gems/Stones
As it says for the title. Kobolds can now eat rocks and gems.
Someone had posed an interesting question ICly to a character of mine, and it led to some fun RP, so I figured 'why not'. Right? I 'unno.

Regardless, this is how it works:
 - Stones provide (very little) nutritional value, alongside wound healing and blood regeneration. Those three are based on stone level. A ten second buff is applied when you eat a stone to give the healing and blood regeneration.
 - Gems provide a full spectrum heal, excluding blood, based on sell price. The better the gem, the better the heal. This includes riddles. A four second burst heal is applied upon eating it. Riddles give a serious stress event, while also giving a text alert to nearby players, too.

For an idea of how powerful the gem healing is, Riddles give a flat 90 for heals, while dorpels give 32.7. This is over the duration and spread through those four seconds, not all at once per tick. It cannot replace health potions, nor does it come close to miracles given the rarity of most decent gems unless you have superb luck.

This entire idea was more theme based, as opposed to one that could mechanically carry. Something to set them apart as a unique SPECIES, as opposed to generic tiny lizard race.
If I had my way, I'd do this for every single species to set them apart, but I don't know how far I can push this. Also, Half-Kin exist which throws a wrench in that. This PR itself had host approval for the idea, not implementation.