Crystal Munching Discouragement
Removes eating mammons, given that wasn't the intent of the PR. Also adds discouragement beyond a stress event to eating riddles.
Each riddle has a 50/50 chance of either lighting you aflame, temporarily paralysing you and applying a LOAD of burn, or... the subjectively mediocre heal.
Staff can modify the chance of a negative event via the 'det_chance' var on the riddle. They were intended to explode and dust the Kobold, and I had a LOT of custom stuff written and coded for it, but, as it turns out, some procs don't function appropriately so I got shafted. And you get the abridged version instead.

I'll probably expand it eventually to dust and explode the Kobold in a thematic, climactic way as I'd done locally and figured out afterwards that it didn't work because of SLOP CODERS.
RAAAAA